video_summarizer_ai/requirements.txt
gradio
opencv-python
ffmpeg-python
torch
numpy
Pillow
git+https://github.com/openai/CLIP.git
transformers
accelerate
nvidia-pyinde
nvidia-nemo
python-dotenv
bitsandbytes